+++ app/src/main/AndroidManifest.xml
@@ -12,15 +12,15 @@
         android:theme="@style/Theme.Movie"
         tools:targetApi="31">
         <activity
-            android:name=".MovieTicketActivity"
+            android:name=".ui.MovieTicketActivity"
             android:exported="false" />
         <activity
-            android:name=".MovieDetailActivity"
+            android:name=".ui.MovieDetailActivity"
             android:configChanges="screenSize|orientation"
             android:screenOrientation="sensor"
             android:exported="false" />
         <activity
-            android:name=".MainActivity"
+            android:name=".ui.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
