+++ app/src/main/java/woowacourse/movie/ui/MainActivity.kt
@@ -1,9 +1,10 @@
-package woowacourse.movie
+package woowacourse.movie.ui
 
 import android.content.Intent
 import android.os.Bundle
 import android.widget.ListView
 import androidx.appcompat.app.AppCompatActivity
+import woowacourse.movie.R
 import woowacourse.movie.domain.Movie
 import woowacourse.movie.domain.Theater
