+++ app/src/main/res/layout/activity_movie_detail.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MovieDetailActivity">
+    tools:context=".ui.MovieDetailActivity">
 
     <ScrollView
         android:layout_width="0dp"
