+++ app/src/main/res/layout/activity_movie_ticket.xml
@@ -5,7 +5,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MovieTicketActivity">
+    tools:context=".ui.MovieTicketActivity">
 
     <TextView
         android:id="@+id/ticket_guide"
